Only build/push bundle image on release
Building/pushing a `latest` bundle image from the main branch makes sense. However, since bundles are immutable once they're release, images should not be overridden on changes to non-latest bundles. This PR removes the workflow that was overriding the image, so that the only time an image is built with a tagged version is when a new release is cut.